Remove kipi-plugins and thinner layers
Most kipi-plugins required a web-browser to be usable, but this image doesn't contain one for the time being. Therefore removing.
Layer installing Digikam was much too big (around 600MB). Losing the connection while downloading that layer is quite disappointing. Therefore shifting the installation of some packages to a different layer.